Disable the syslog reader
It gets really slow and doesn't work properly when the log file
has <ins> in it, and a binary index section at the end. It also
seems like the file sometimes is opened with a delete mode, which
conflicts with the log reader, and causes the log to not be written.